kicking around some refactoring.
