+++ typescript/FlareClusterLayer_v4.ts
@@ -2,6 +2,8 @@
 
 
 import * as GraphicsLayer from "esri/layers/GraphicsLayer";
+import * as FeatureLayer from "esri/layers/FeatureLayer";
+import * as GroupLayer from "esri/layers/GroupLayer";
 import * as ClassBreaksRenderer from "esri/renderers/ClassBreaksRenderer";
 import * as PopupTemplate from "esri/PopupTemplate";
 import * as SimpleMarkerSymbol from "esri/symbols/SimpleMarkerSymbol";
@@ -29,9 +31,9 @@ import * as domConstruct from 'dojo/dom-construct';
 import * as query from 'dojo/query';
 import * as domAttr from 'dojo/dom-attr';
 import * as domStyle from 'dojo/dom-style';
+import * as deferred from 'dojo/Deferred';
 
-
-interface FlareClusterLayerProperties extends __esri.GraphicsLayerProperties {
+interface FlareClusterLayerProperties extends __esri.GroupLayerProperties {
 
     singleRenderer?: any;
     renderer?: ClassBreaksRenderer;
@@ -69,19 +71,19 @@ interface FlareClusterLayerProperties extends __esri.GraphicsLayerProperties {
 
 //This is how you have to extend classes in arcgis api that are a subclass of Accessor.
 //Will likely change in future releases. See these links - https://github.com/Esri/jsapi-resources/issues/40 & https://github.com/ycabon/extend-accessor-example
-interface BaseGraphicsLayer extends GraphicsLayer { }
-interface BaseGraphicsLayerConstructor { new (options?: __esri.GraphicsLayerProperties): BaseGraphicsLayer; }
-let baseGraphicsLayer: BaseGraphicsLayerConstructor = accessorSupportDecorators.declared(<any>GraphicsLayer);
+interface BaseGroupLayer extends GroupLayer { }
+interface BaseGroupLayerConstructor { new (options?: __esri.GraphicsLayerProperties): BaseGroupLayer; }
+let baseGroupLayer: BaseGroupLayerConstructor = accessorSupportDecorators.declared(<any>GroupLayer);
 
 @accessorSupportDecorators.subclass("FlareClusterLayer")
-export class FlareClusterLayer extends baseGraphicsLayer {
+export class FlareClusterLayer extends baseGroupLayer {
 
     singlePopupTemplate: PopupTemplate;
     singleRenderer: any;
     renderer: ClassBreaksRenderer;
     areaRenderer: ClassBreaksRenderer;
     spatialReference: SpatialReference;
-    refreshOnStationary: boolean;
+    refreshOnStationary: boolean = true;
     gridClusters: GridCluster[];
     clusterRatio: number;
     clusterToScale: number;
@@ -95,23 +97,27 @@ export class FlareClusterLayer extends baseGraphicsLayer {
     flareTextSymbol: TextSymbol;
     flareShowMode: string;
     xPropertyName: string;
-    yPropertyName: string;
+    yPropertyName: string; 
     zPropertyName: string;
     displaySubTypeFlares: boolean;
     subTypeFlareProperty: string;
 
-    filters: PointFilter[];
+    private _data: any[];
+    private _singleLayer: GraphicsLayer;
+    private _clusterLayer: GraphicsLayer;
+    private _areaLayer: GraphicsLayer;
+    private _flareLayer: GraphicsLayer;
+    private _isClustered: boolean;
+    private _activeView: View;
+    private _viewLoadCount: number = 0;
 
-    activeView: any;
+    private get _currentExtent(): Extent {
+        return this._activeView ? this._activeView["extent"] : undefined;
+    }
 
-    private isClustered: boolean;
-    private readyToDraw: boolean;
-    private queuedInitialDraw: boolean;
-    private data: any[];
-    private layerViews: any[];
-
-    private viewLoadCount;
-    private viewPopupMessageEnabled;
+    private get _currentScale(): number {
+        return this._activeView ? this._activeView["scale"] : undefined;
+    }
 
     constructor(options: FlareClusterLayerProperties) {
 
@@ -121,10 +127,6 @@ export class FlareClusterLayer extends baseGraphicsLayer {
         if (!options) {
             options = {};
         }
-
-        this.layerViews = [];
-        this.viewLoadCount = 0;
-        this.viewPopupMessageEnabled = true;
 
         this.singlePopupTemplate = options.singlePopupTemplate;
         this.clusterRatio = options.clusterRatio || 75;
@@ -143,8 +145,6 @@ export class FlareClusterLayer extends baseGraphicsLayer {
         this.xPropertyName = options.xPropertyName || "x";
         this.yPropertyName = options.yPropertyName || "y";
         this.zPropertyName = options.zPropertyName || "z";
-
-        this.filters = options.filters || [];
 
         this.displaySubTypeFlares = options.displaySubTypeFlares === true;
         this.subTypeFlareProperty = options.subTypeFlareProperty || undefined;
@@ -174,25 +174,27 @@ export class FlareClusterLayer extends baseGraphicsLayer {
             }
         });
 
-        this.data = options.data || undefined;
-        this.on("layerview-create", (evt) => this.layerViewCreated(evt));
+        this._data = options.data || undefined;
 
-        if (this.data) {
-            this.drawData();
+        this._setupLayers();
+
+        this.on("layerview-create", (evt) => this._layerViewCreated(evt));
+
+        if (this._data) {
+            //this.drawData();
         }
     }
-     
 
-    private layerViewCreated(evt) {
+    private _layerViewCreated(evt) {
 
         if (evt.layerView.view.type === "2d") {
             //this is map view so set up a watch to find out when the vector group has been created. Since 4.1 - use the gfx poroperty pm the graphicsView instead of group??
-            watchUtils.whenDefinedOnce(evt.layerView._graphicsView, "gfx", (vectorGroup, b, c, graphicsView) => this.vectorGroupCreated(vectorGroup, b, c, graphicsView));
+            //watchUtils.whenDefinedOnce(evt.layerView._graphicsView, "gfx", (vectorGroup, b, c, graphicsView) => this.vectorGroupCreated(vectorGroup, b, c, graphicsView));
         }
         else {
             //this is 3d so add a custom external rendeder to hook into webgl pipeline to do things.
-            let fclExternalRenderer = new FlareClusterLayerExternalRenderer(evt.layerView);
-            externalRenderers.add(evt.layerView.view, fclExternalRenderer);
+            //let fclExternalRenderer = new FlareClusterLayerExternalRenderer(evt.layerView);
+            //externalRenderers.add(evt.layerView.view, fclExternalRenderer);
         }
 
         //Refresh the data when the view is stationary if not set to false in options.
@@ -200,108 +202,43 @@ export class FlareClusterLayer extends baseGraphicsLayer {
             watchUtils.pausable(evt.layerView.view, "stationary", (isStationary, b, c, view) => this.viewStationary(isStationary, b, c, view));
         }
 
-        this.viewPopupMessageEnabled = evt.layerView.view.popup.messageEnabled;
+        //this.viewPopupMessageEnabled = evt.layerView.view.popup.messageEnabled;
 
         //watch this property so we can not display popups for graphics we don't want to.
-        watchUtils.watch(evt.layerView.view.popup.viewModel, "selectedFeature", (selectedFeature, b, c, viewModel) => this.viewPopupSelectedFeatureChange(selectedFeature, b, c, viewModel));
+        //watchUtils.watch(evt.layerView.view.popup.viewModel, "selectedFeature", (selectedFeature, b, c, viewModel) => this.viewPopupSelectedFeatureChange(selectedFeature, b, c, viewModel));
 
-        this.layerViews.push(evt.layerView);
+        //this.layerViews.push(evt.layerView);
 
-        if (this.viewLoadCount === 0) {
-            this.activeView = evt.layerView.view;
+        if (this._viewLoadCount === 0) {
+            this._activeView = evt.layerView.view;
         }
-        this.viewLoadCount++;
-    }
-
-    private viewPopupSelectedFeatureChange(selectedFeature, b, c, viewModel) {
-        //There has got to be an better way to not show popups for certain graphics!
-
-        if (!selectedFeature) {
-            //reset the popup message for the view so this layer doens't affect other layers.
-            viewModel.view.popup.messageEnabled = this.viewPopupMessageEnabled;
-            return;
-        }
-
-        //if this is a cluster type graphic then hide the popup
-        if (selectedFeature.attributes.isFlare || selectedFeature.attributes.isCluster || selectedFeature.attributes.isClusterArea) {
-            viewModel.features = [];
-            viewModel.view.popup.messageEnabled = false;
-            viewModel.view.popup.close();
-        }
-
-    }
-
-    private viewStationary(isStationary, b, c, view) {
-        if (this.data && isStationary) {
-            this.activeView = view;
-            this.drawData();
-        }
-    }
-
-    private vectorGroupCreated(vectorGroup, b, c, graphicsView) {
-
-        //need to set the z-index of the layer view container to something higher than the basemap container so it will receive events. This is a hack I think.
-        var parentDiv = graphicsView.gfx._parent;
-        var newStyle = parentDiv.getAttribute("style") + ";z-index:10";
-        parentDiv.setAttribute("style", newStyle);
-
-        let existingGroup = graphicsView.group;
-        graphicsView.group = new FlareClusterVectorGroup({
-            view: existingGroup.view,
-            x: existingGroup.x,
-            y: existingGroup.y,
-            resolution: existingGroup.resolution,
-            rotation: existingGroup.rotation,
-            surface: existingGroup.surface,
-            layer: this
-        });
-
-        this.readyToDraw = true;
-        if (this.queuedInitialDraw) {
-            this.drawData(existingGroup.view);
-            this.queuedInitialDraw = false;
-        }
-    }
-
-    clear() {
-        this.removeAll();
-        for (let lv of this.layerViews) {
-            if (lv._graphicsView && lv._graphicsView.group) {
-                //this is a 2d layer view so clear the vector groups vector array
-                lv._graphicsView.group.vectors = [];
-            }
-        }
+        this._viewLoadCount++;
     }
 
     setData(data: any[], drawData: boolean = true) {
-
-        this.data = data;
+        this._data = data;
         if (drawData) {
             this.drawData();
         }
     }
 
- 
+
     drawData(activeView?: any) {
-        //Not ready to draw yet so queue one up
-        if (!this.readyToDraw) {
-            this.queuedInitialDraw = true;
-            return;
-        }
-
+        
         if (activeView) {
-            this.activeView = activeView;
+            this._activeView = activeView;
         }
 
-        //no data set and no active (visible) view found so return
-        if (!this.data || !this.activeView) return;
+        console.log('in draw data');
+        if(!this._activeView || !this._data) return;
 
         this.clear();
         console.time("draw-data");
 
-        this.isClustered = this.clusterToScale < this.activeView["scale"];
+        this._isClustered = this.clusterToScale < this._currentScale;
 
-        console.log("draw data " + this.activeView.type);
+        //console.log("draw data " + this.activeView.type);
+        console.log("draw data ");
 
         let graphics: Graphic[] = [];
 
@@ -309,7 +246,7 @@ export class FlareClusterLayer extends baseGraphicsLayer {
         //The webextent will need to be normalized since panning over the international dateline will cause
         //cause the extent to shift outside the -180 to 180 degree window.  If we don't normalize then the
         //clusters will not be drawn if the map pans over the international dateline.
-        let webExtent = !this.activeView.extent.spatialReference.isWebMercator ? webMercatorUtils.project(this.activeView.extent, new SpatialReference({ "wkid": 102100 })) : this.activeView.extent;
+        let webExtent: any = !this._currentExtent.spatialReference.isWebMercator ? <Extent>webMercatorUtils.project(this._currentExtent, new SpatialReference({ "wkid": 102100 })) : this._currentExtent;
         let extentIsUnioned = false;
         
         let normalizedWebExtent = webExtent.normalize();
@@ -319,19 +256,19 @@ export class FlareClusterLayer extends baseGraphicsLayer {
             extentIsUnioned = true;
         }
 
-        if (this.isClustered) {
-            this.createClusterGrid(webExtent, extentIsUnioned);
+        if (this._isClustered) {
+            this._createClusterGrid(webExtent, extentIsUnioned);
         }
 
 
-        let web: number[], obj: any, dataLength = this.data.length, xVal: number, yVal: number;
+        let web: number[], obj: any, dataLength = this._data.length, xVal: number, yVal: number;
         for (let i = 0; i < dataLength; i++) {
-            obj = this.data[i];
+            obj = this._data[i];
 
             //check if filters are specified and continue if this object doesn't pass
-            if (!this.passesFilter(obj)) {
-                continue;
-            }
+            //if (!this.passesFilter(obj)) {
+            //    continue;
+            //}
 
             xVal = obj[this.xPropertyName];
             yVal = obj[this.yPropertyName];
@@ -348,7 +285,7 @@ export class FlareClusterLayer extends baseGraphicsLayer {
                 continue;
             }
 
-            if (this.isClustered) {
+            if (this._isClustered) {
 
                 //loop cluster grid to see if it should be added to one
                 for (let j = 0, jLen = this.gridClusters.length; j < jLen; j++) {
@@ -395,19 +332,19 @@ export class FlareClusterLayer extends baseGraphicsLayer {
             }
             else {
                 //not clustered so just add every obj
-                this.createSingle(obj);
+                //this.createSingle(obj);
             }
         }
 
-        if (this.isClustered) {
+        if (this._isClustered) {
             for (let i = 0, len = this.gridClusters.length; i < len; i++) {
                 if (this.gridClusters[i].clusterCount < this.clusterMinCount) {
                     for (let j = 0, jlen = this.gridClusters[i].singles.length; j < jlen; j++) {
-                        this.createSingle(this.gridClusters[i].singles[j]);
+                        //this.createSingle(this.gridClusters[i].singles[j]);
                     }
                 }
                 else if (this.gridClusters[i].clusterCount > 1) {
-                    this.createCluster(this.gridClusters[i]);
+                    //this.createCluster(this.gridClusters[i]);
                 }
             }
         }
@@ -418,251 +355,27 @@ export class FlareClusterLayer extends baseGraphicsLayer {
         console.timeEnd("draw-data");
     }
 
-
-    private passesFilter(obj: any): boolean {
-        if (!this.filters || this.filters.length === 0) return true;
-
-        for (let i = 0, len = this.filters.length; i < len; i++) {
-            let filter = this.filters[i];
-            if (obj[filter.propertyName] == null) continue;
-            if (filter.propertyValues.indexOf(obj[filter.propertyName]) !== -1) {
-                return false; //return false as the object contains a property with this filters name and the value is contained within the filters value array.
-            }
-        }
-
-        return true;
-    }
-
-    private createSingle(obj) {
-        let point = new Point({
-            x: obj[this.xPropertyName], y: obj[this.yPropertyName], z: obj[this.zPropertyName]
-        });
-
-        if (!point.spatialReference.isWebMercator) {
-            point = <Point>webMercatorUtils.geographicToWebMercator(point);
-        }
-
-        let graphic = new Graphic({
-            geometry: point,
-            attributes: obj
-        });
-
-        graphic.popupTemplate = this.singlePopupTemplate;
-        if (this.singleRenderer) {
-            let symbol = this.singleRenderer.getSymbol(graphic, this.activeView);
-            graphic.symbol = symbol;
-
-        }
-
-        this.add(graphic);
+    
+    clear() {
+        this._singleLayer.removeAll();
+        this._clusterLayer.removeAll();
+        this._areaLayer.removeAll();
+        this._flareLayer.removeAll();
 
     }
 
-
-    private createCluster(cluster: GridCluster) {
-
-        //make sure all geometries added to Graphic objects are in web mercator otherwise wrap around doesn't work.
-        let point = new Point({ x: cluster.x, y: cluster.y });
-        if (!point.spatialReference.isWebMercator) {
-            point = <Point>webMercatorUtils.geographicToWebMercator(point);
+     private viewStationary(isStationary, b, c, view) {
+        if (this._data && isStationary) {
+            this._activeView = view;
+            this.drawData();
         }
-
-        let attributes: any = {
-            x: cluster.x,
-            y: cluster.y,
-            clusterCount: cluster.clusterCount,
-            isCluster: true,
-            clusterObject: cluster
-        }
-
-        let graphic = new Graphic({
-            attributes: attributes,
-            geometry: point
-        });
-
-        graphic.popupTemplate = null;
-
-        this.add(graphic);
-
-        //also create a text symbol to display the cluster count
-        let textSymbol = this.textSymbol.clone();
-        textSymbol.text = cluster.clusterCount.toString();
-
-        let tg = new Graphic({
-            geometry: point,
-            attributes: {
-                isClusterText: true,
-                clusterGraphicId: graphic["uid"]
-            },
-            symbol: textSymbol
-        });
-        tg.popupTemplate = null;
-        this.add(tg);
-        graphic.attributes.textGraphic = tg;
-        
-
-        //add an area graphic to display the bounds of the cluster if configured to
-        let areaGraphic;
-        if (this.clusterAreaDisplay && cluster.points && cluster.points.length > 0) {
-            if (!this.areaRenderer) {
-                console.error("_createCluster: areaRenderer must be set if clusterAreaDisplay is set.");
-                return;
-            }
-
-            let mp = new Multipoint();
-            mp.points = cluster.points;
-            let area: any = geometryEngine.convexHull(mp, true); //use convex hull on the points to get the boundary
-            
-            let areaAttr: any = {
-                x: cluster.x,
-                y: cluster.y,
-                clusterCount: cluster.clusterCount,
-                clusterGraphicId: graphic["uid"],
-                isClusterArea: true
-            }
-
-            if (area.rings && area.rings.length > 0) {
-                let areaPoly = new Polygon(); //had to create a new polygon and fill it with the ring of the calculated area for SceneView to work.
-                areaPoly = areaPoly.addRing(area.rings[0]);
-
-                if (!areaPoly.spatialReference.isWebMercator) {
-                    areaPoly = <Polygon>webMercatorUtils.geographicToWebMercator(areaPoly);
-                }
-
-                areaGraphic = new Graphic({ geometry: areaPoly, attributes: areaAttr });
-                let areaInfo = this.areaRenderer.getClassBreakInfo(areaGraphic);
-                areaGraphic.symbol = areaInfo.symbol;
-                areaGraphic.cluster = cluster;
-                areaGraphic.popupTemplate = null;
-                this.add(areaGraphic);
-
-                graphic.attributes.areaGraphic = areaGraphic;
-            }
-        }
-
-        //create the graphics for the flares and assign to attribute
-        let flareGraphics = this.createFlares(cluster, graphic);
-        graphic.attributes.flareGraphics = flareGraphics;
-
     }
 
-    private createFlares(clusterObject: GridCluster, clusterGraphic: Graphic) {
-
-        let flareGraphics: Graphic[] = [];
-
-        //check if we need to create flares for the cluster
-        let singleFlares = (clusterObject.singles && clusterObject.singles.length > 0) && (clusterObject.clusterCount <= this.maxSingleFlareCount);
-        let subTypeFlares = !singleFlares && (clusterObject.subTypeCounts && clusterObject.subTypeCounts.length > 0);
-
-        if (!singleFlares && !subTypeFlares) {
-            return flareGraphics; //no flares required
-        }
-
-        let flareObjects = [];
-        if (singleFlares) {
-            for (var i = 0, len = clusterObject.singles.length; i < len; i++) {
-                flareObjects.push({
-                    tooltipText: clusterObject.singles[i][this.singleFlareTooltipProperty],
-                    flareText: "",
-                    singleData: clusterObject.singles[i]
-                });
-            }
-        }
-        else if (subTypeFlares) {
-
-            //sort sub types by highest count first
-            var subTypes = clusterObject.subTypeCounts.sort(function (a, b) {
-                return b.count - a.count;
-            });
-
-            for (var i = 0, len = subTypes.length; i < len; i++) {
-                flareObjects.push({
-                    tooltipText: subTypes[i].count + " - " + subTypes[i].name,
-                    flareText: subTypes[i].count
-                });
-            }
-        }
-
-        //if there are more flare objects to create that the maxFlareCount and this is a one of those - create a summary flare that contains '...' as the text and make this one part of it 
-        let willContainSummaryFlare = flareObjects.length > this.maxFlareCount;
-        let flareCount = willContainSummaryFlare ? this.maxFlareCount : flareObjects.length;
-
-        for (let i = 0; i < flareCount; i++) {
-
-            //exit if we've hit the maxFlareCount - a summary would have been created on the last one
-            if (i >= this.maxFlareCount) {
-                break;
-            }
-
-            let fo = flareObjects[i];
-
-            //set some attribute data
-            let flareAttributes = {
-                isFlare: true,
-                isSummaryFlare: false,
-                tooltipText: "",
-                flareTextGraphic: undefined,
-                clusterGraphicId: clusterGraphic["uid"]
-            };
-
-            let flareTextAttributes = {};
-
-            //Do a couple of things differently if this is a summary flare or not
-            let tooltipText = "";
-            let isSummaryFlare = willContainSummaryFlare && i >= this.maxFlareCount - 1;
-            if (isSummaryFlare) {
-
-                flareAttributes.isSummaryFlare = true;
-
-                //multiline tooltip for summary flares, ie: greater than this.maxFlareCount flares per cluster
-                for (let j = this.maxFlareCount - 1, jlen = flareObjects.length; j < jlen; j++) {
-                    tooltipText += j > (this.maxFlareCount - 1) ? "\n" : "";
-                    tooltipText += flareObjects[j].tooltipText;
-                }
-            }
-            else {
-                tooltipText = fo.tooltipText;
-            }
-
-            flareAttributes.tooltipText = tooltipText;
-
-            //create a graphic for the flare and for the flare text, don't worry about positioning though, it will be set in the view specific code
-            let flareGraphic = new Graphic({
-                attributes: flareAttributes,
-                symbol: this.flareSymbol,
-                geometry: clusterGraphic.geometry, //default geometry to be cluster point
-                popupTemplate: null
-            });
-            flareGraphics.push(flareGraphic);
-
-            if (fo.flareText) {
-                let textSymbol = this.flareTextSymbol.clone();
-                textSymbol.text = !isSummaryFlare ? fo.flareText.toString() : "...";
-                let flareTextGraphic = new Graphic({
-                    attributes: {
-                        isFlareText: true,
-                        clusterGraphicId: clusterGraphic["id"],
-                        flareGraphicId: flareGraphic["id"]
-                    },
-                    symbol: textSymbol,
-                    geometry: clusterGraphic.geometry //default geometry to be cluster point
-                });
-
-                //add text to attributes of flare graphic
-                flareGraphic.attributes.flareTextGraphic = flareTextGraphic;
-                flareGraphics.push(flareTextGraphic);
-            }
-        }
-
-        this.addMany(flareGraphics);
-        return flareGraphics;
-    }
-
-    private createClusterGrid(webExtent: Extent, extentIsUnioned: boolean) {
+    private _createClusterGrid(webExtent: Extent, extentIsUnioned: boolean) {
 
         //get the total amount of grid spaces based on the height and width of the map (divide it by clusterRatio) - then get the degrees for x and y 
-        var xCount = Math.round(this.activeView.width / this.clusterRatio);
-        var yCount = Math.round(this.activeView.height / this.clusterRatio);
+        var xCount = Math.round(this._activeView.width / this.clusterRatio);
+        var yCount = Math.round(this._activeView.height / this.clusterRatio);
 
         //if the extent has been unioned due to normalization, double the count of x in the cluster grid as the unioning will halve it.
         if (extentIsUnioned) {
@@ -695,7 +408,27 @@ export class FlareClusterLayer extends baseGraphicsLayer {
             }
         }
     }
+
+    /**
+     * Setup the layers.
+     */
+    _setupLayers() {
+        this._singleLayer = new GraphicsLayer();
+        this._clusterLayer = new GraphicsLayer();
+        this._areaLayer = new GraphicsLayer();
+        this._flareLayer = new GraphicsLayer();
+         
+        this.addMany([this._areaLayer, this._clusterLayer, this._singleLayer, this._flareLayer]);        
+
+        this._clusterLayer["popupEnabled"] = false;
+        this._areaLayer["popupEnabled"] = false;
+        this._flareLayer["popupEnabled"] = false;
+
+    }
+
+    
 }
+
 
 class GridCluster {
     extent: any;
@@ -714,877 +447,6 @@ export class PointFilter {
     constructor(name?: string, values?: any[]) {
         this.propertyName = name;
         this.propertyValues = values;
-    }
-
-}
-
-class FlareClusterVectorGroup extends VectorGroup {
-
-    drawComplete: any;
-    eventHandles: any[] = [];
-    animationsRunning: any[] = [];
-    activeClusterVector: any;
-    layer: any;
-    view: any;
-
-    private clusterVectors: any[] = [];
-
-    constructor(options: any) {
-        super(options);
-    }
-
-    removeVector(a) {
-        if (!a) return;
-        if (a.clusterGroup) {
-            domConstruct.destroy(a.clusterGroup.rawNode);
-        }
-        else if (a.shape) {
-            domConstruct.destroy(a.shape.rawNode);
-        }
-        return super.removeVector(a);
-
-    }
-
-    draw(a) {
-
-        //only applies to 2d and only if there's vectors to draw
-        if (this.layer.activeView.type !== "2d" || this.vectors.length === 0) {
-            return;
-        }
-
-        //destroy all cluster objects
-        query(".cluster-object", this.surface.rawNode).forEach(domConstruct.destroy);
-
-        this.clusterVectors = [];
-
-        this.transform || this._updateTransform();
-        let view = this.view, b = view.state;
-        if (0 < b.width - b.worldScreenWidth) {
-            var d = b.size.concat();
-            d[0] = Math.min(d[0], Math.round(b.worldScreenWidth));
-            this._viewExtent = viewpointUtils.getExtent(new Extent, b.viewpoint, d)
-        } else
-            this._viewExtent = view.extent;
-        this._viewGeoExtent = view.spatialReference.isWebMercator ? <Extent>webMercatorUtils.webMercatorToGeographic(this._viewExtent) : null;
-        this.surface.openBatch();
-
-        let index = 0; 
-        let c;
-        for (d = this.vectors.length; index < d; index++) {
-            (c = this.vectors[index]) && super.drawVector(c, a);
-
-            let v = this.vectors[index];
-            if (!v.shape) continue;
-
-            if (v.graphic.attributes.isCluster) {
-                v.shape.rawNode.setAttribute("class", "cluster"); //this is a cluster
-
-                v.flareVectors = [];
-                v.flareTextVectors = [];
-                this.clusterVectors.push(v);
-            }
-            else if (v.graphic.attributes.isClusterText) {
-                //this is cluster text
-                v.shape.rawNode.setAttribute("class", "cluster-text");
-                v.shape.rawNode.setAttribute("pointer-events", "none");
-
-                //assign to a property on the cluster vector
-                let cv = this.getClusterVectorByGraphicId(v.graphic.attributes.clusterGraphicId);
-                if (cv) {
-                    cv.textVector = v;
-                }
-            }
-            else if (v.graphic.attributes.isClusterArea) {
-                v.shape.rawNode.setAttribute("class", "cluster-area");
-                
-                if (v.shape.rawNode.parentNode) {
-                    v.shape.moveToBack();
-                }
-                
-                if (this.layer.clusterAreaDisplay === "activated") {
-                    //remove the node from the dom (try and keep it as light as possible)
-                    this.removeNodeFromDom(v.shape.rawNode);
-                }
-                else if (this.layer.clusterAreaDisplay === "always") {
-                    v.shape.rawNode.setAttribute("class", "cluster-area activated"); //add the activated class if the area is always visible                
-                }
-
-                //assign to a property on the cluster vector
-                let cv = this.getClusterVectorByGraphicId(v.graphic.attributes.clusterGraphicId);
-                if (cv) {
-                    cv.areaVector = v;
-                }
-
-            }
-            else if (v.graphic.attributes.isFlare) {
-                //remove the node from the dom (try and keep it as light as possible)
-                this.removeNodeFromDom(v.shape.rawNode);
-
-                //assign to a property on the cluster vector
-                let cv = this.getClusterVectorByGraphicId(v.graphic.attributes.clusterGraphicId);
-                if (cv) {
-                    cv.flareVectors.push(v);
-                }
-            }
-            else if (v.graphic.attributes.isFlareText) {
-                v.shape.rawNode.setAttribute("pointer-events", "none");
-
-                //remove the node from the dom (try and keep it as light as possible)
-                this.removeNodeFromDom(v.shape.rawNode);
-
-                //assign to a property on the flare shape vector - the flare shape vector should be the previous entry in the array
-                let flareShapeVector = this.vectors[index - 1];
-                if (flareShapeVector.graphic.attributes.isFlare) {
-                    flareShapeVector.textVector = v;
-                }
-            }
-        }
-
-        this.surface.closeBatch()
-
-        if (!this.layer.clusterAreaDisplay) {
-            //area should not be displayed at all so destroy the nodes
-            query(".cluster-area", this.surface.rawNode).forEach(domConstruct.destroy);
-        }
-
-        //remove any event handlers added previously
-        for (let evtHandle of this.eventHandles) {
-            evtHandle.remove();
-        }
-
-        for (let cv of this.clusterVectors) {
-            this.initCluster(cv);
-        }
-
-        let clusterGroups = query(".cluster-group", this.surface.rawNode);
-
-        this.eventHandles.push(on.pausable(clusterGroups, "mouseenter", (evt) => this.clusterMouseEnter(evt)));
-        this.eventHandles.push(on.pausable(clusterGroups, "mouseleave", (evt) => this.clusterMouseLeave(evt)));
-    }
-
-    private initCluster(clusterVector) {
-
-        if (!clusterVector.shape || !clusterVector.textVector.shape) {
-            return;
-        }
-
-        //for each clusterVector create an svg group that contains the cluster circle and the cluster text shapes.
-        var group = this.surface.createGroup();
-        group.rawNode.setAttribute("class", "cluster-group");
-        group.add(clusterVector.shape);
-        group.add(clusterVector.textVector.shape);
-
-        this.addClassToNode(group.rawNode, "created", 5);
-
-        clusterVector.clusterGroup = group;
-    }
-
-
-    private clusterMouseEnter(evt) {
-
-        //get the cluster vector from the vectors array
-        let vector = this.getClusterVectorFromGroupNode(evt.target);
-        this.activateCluster(vector);
-    }
-
-    private activateCluster(vector) {
-
-        if (vector === this.activeClusterVector) return; //already active
-
-        if (this.activeClusterVector) {
-            this.deactivateCluster(this.activeClusterVector);
-        }
-
-        this.addClassToNode(vector.clusterGroup.rawNode, "activated", 5);
-
-        let center = this.getShapeCenter(vector.shape);
-        vector.center = center;
-
-        //Handle scaling and moving to front as well.
-        vector.clusterGroup.moveToFront();
-        let scaleAnims: any[] = [];
-
-        if (vector.areaVector && vector.areaVector.shape && this.layer.clusterAreaDisplay === "activated") {
-            //add the area vector shape into the dom and add an activated class
-            this.surface.rawNode.appendChild(vector.areaVector.shape.rawNode);
-            vector.areaVector.shape.moveToBack();
-            this.addClassToNode(vector.areaVector.shape.rawNode, "activated", 5);
-        }
-
-        this.createFlares(vector);
-        this.activeClusterVector = vector;
-
-    }
-
-
-    private clusterMouseLeave(evt) {
-        let vector = this.getClusterVectorFromGroupNode(evt.target);
-        this.deactivateCluster(vector);
-    }
-
-    private deactivateCluster(vector) {
-        let center = vector.center;
-
-        //remove any flare shapes from the DOM
-        for (let i = 0, len = vector.flareVectors.length; i < len; i++) {
-            this.removeClassFromNode(vector.flareVectors[i].shape.rawNode, "activated", 5);
-            this.removeNodeFromDom(vector.flareVectors[i].shape.rawNode);
-        }
-
-        //destroy all flare objects in this cluster group
-        query(".cluster-object", this.surface.rawNode).forEach(domConstruct.destroy);
-        this.removeClassFromNode(vector.clusterGroup.rawNode, "activated", 5);
-
-        if (vector.areaVector && vector.areaVector.shape && this.layer.clusterAreaDisplay === "activated") {
-            //remove the area vector from the dom
-            this.removeClassFromNode(vector.areaVector.shape.rawNode, "activated", 5);
-            this.removeNodeFromDom(vector.areaVector.shape.rawNode);
-        }
-
-        this.activeClusterVector = null;
-    }
-
-    private fixMouseEnter(evt) {
-        this.surface.rawNode.style.cursor = "pointer";
-    }
-
-    private fixMouseLeave(evt) {
-        this.surface.rawNode.style.cursor = "default";
-    }
-
-
-    private createFlares(clusterVector) {
-
-        let flareVectors = clusterVector.flareVectors;
-        if (!flareVectors || flareVectors.length === 0) return;
-
-        let flareCount = flareVectors.length;
-
-        //create and add a graphic to represent the flare circle
-        let bbox = clusterVector.shape.getBoundingBox();
-        let radius = 8;
-        let buffer = 8;
-
-        let clusterGroup = clusterVector.clusterGroup;
-
-        //create a transparent circle that contains the boundary of the flares, this is to make sure the mouse events don't fire moving in between flares
-        let conCircleRadius = (clusterVector.center.x - (bbox.x - radius - buffer)) + radius; //get the radius of the circle to contain everything
-        let containerCircle = clusterGroup.createCircle({ cx: clusterVector.center.x, cy: clusterVector.center.y, r: conCircleRadius })
-            .setFill([0, 0, 0, 0]);
-        containerCircle.rawNode.setAttribute("class", "flare-container cluster-object");
-
-        //if there's an even amount of flares, position the first flare to the left, minus 180 from degree to do this.
-        //for an add amount position the first flare on top, -90 to do this. Looks more symmetrical this way.
-        let degreeVariance = (flareCount % 2 === 0) ? -180 : -90;
-        let viewRotation = this.view.rotation;
-
-        //array to hold the animations for displaying flares
-        let stAnims = [];
-        for (let i = 0; i < flareCount; i++) {
-
-            let flareVector = flareVectors[i];
-
-            //get the position of the flare to be placed around the container circle.
-            let degree = parseInt(((360 / flareCount) * i).toFixed());
-            degree = degree + degreeVariance;
-
-            //take into account any rotation on the view
-            if (viewRotation !== 0) {
-                degree -= viewRotation;
-            }
-
-            var radian = degree * (Math.PI / 180);
-
-            //calc the center point of the flare
-            let screenPoint = {
-                x: clusterVector.center.x + (conCircleRadius - radius - 5) * Math.cos(radian),
-                y: clusterVector.center.y + (conCircleRadius - radius - 5) * Math.sin(radian)
-            };
-
-            //create a group to hold the flare an possibly the text for the flare
-            let flareGroup = clusterGroup.createGroup({ x: screenPoint.x, y: screenPoint.y });
-            flareGroup.rawNode.setAttribute("class", "cluster-object flare-group");
-            flareGroup.add(flareVector.shape);
-
-            //if this flare hasn't had it's position set, set it now. Transforming the exsiting location - cluster location to the actual screen location using dx and dy on the shape.
-            if (!flareVector.shape.positionSet) {
-                let transform = flareVector.shape.getTransform();
-                let flareCenter = this.getShapeCenter(flareVector.shape);
-                let diff = {
-                    x: screenPoint.x - flareCenter.x,
-                    y: screenPoint.y - flareCenter.y
-                };
-
-                transform.dx = transform.dx + diff.x;
-                transform.dy = transform.dy + diff.y;
-
-                flareVector.shape.setTransform(transform);
-                flareVector.shape.positionSet = true;
-            }
-
-            //if this flare has a text graphic add it and position as well
-            if (flareVector.textVector) {
-                flareGroup.add(flareVector.textVector.shape);
-                if (!flareVector.textVector.shape.positionSet) {
-                    let textTransform = flareVector.textVector.shape.getTransform();
-                    let flareCenter = this.getShapeCenter(flareVector.textVector.shape);
-                    let diff = {
-                        x: screenPoint.x - flareCenter.x,
-                        y: screenPoint.y - flareCenter.y
-                    };
-
-                    textTransform.dx = textTransform.dx + diff.x;
-                    textTransform.dy = textTransform.dy + diff.y;
-                    flareVector.textVector.shape.setTransform(textTransform);
-                    flareVector.textVector.shape.positionSet = true;
-                }
-            }
-
-            //set the group to be scaled to 0 by default.
-            flareGroup.rawNode.setAttribute("data-tooltip", flareVector.graphic.attributes.tooltipText);
-            flareGroup.rawNode.setAttribute("data-center-x", screenPoint.x);
-            flareGroup.rawNode.setAttribute("data-center-y", screenPoint.y);
-            flareGroup.isSummaryFlare = flareVector.graphic.attributes.isSummaryFlare;
-
-            //add activated class for styling/transitions
-            this.addClassToNode(flareGroup.rawNode, "activated", 10);
-
-            flareGroup.mouseEnter = on.pausable(flareGroup.rawNode, "mouseenter", (e) => this.createTooltip(e));
-            flareGroup.mouseLeave = on.pausable(flareGroup.rawNode, "mouseleave", (e) => this.destroyTooltip(e));
-
-        }
-
-    }
-
-    private createTooltip(e) {
-        if (!this.layer) return;
-
-        let flareGroupNode = e.gfxTarget ? e.gfxTarget.rawNode : e.target;
-        let shape = flareGroupNode.__gfxObject__;
-
-        //destory any existing tooltips.
-        query(".tooltip-text", this.surface.rawNode).forEach(domConstruct.destroy);
-
-        let tooltipLength = query(".tooltip-text", shape.rawNode).length;
-        if (tooltipLength > 0) {
-            return;
-        }
-
-        //get the text from the data-tooltip attribute of the shape object
-        let text = shape.rawNode.getAttribute("data-tooltip");
-        if (!text) {
-            console.log("no data-tooltip attribute on element");
-            return;
-        }
-
-        //split on /n character that should be in tooltip to signify multiple lines
-        let lines = text.split("\n");
-
-        //read the center positions from the shape, attributes must be set on whatever node is being passed in. Calculating from getboundingBox wasn't working for some reason
-        let xPos = parseInt(shape.rawNode.getAttribute("data-center-x"));
-
-        //align on top for normal flare, align on bottom for summary flares.
-        let centerY = parseInt(shape.rawNode.getAttribute("data-center-y"));
-        let yPos = !shape.isSummaryFlare ? centerY - 12 : centerY + 17;
-
-        //create a group to hold the tooltip elements
-        let tooltipGroup = shape.createGroup({ x: xPos, y: yPos });
-        tooltipGroup.rawNode.setAttribute("class", "tooltip-text");
-
-
-        let textShapes = [];
-        for (let i = 0, len = lines.length; i < len; i++) {
-            let textShape = tooltipGroup.createText({ x: xPos, y: yPos + (i * 10), text: lines[i], align: 'middle' })
-                .setFill("#000")
-                .setFont({ size: 8, family: this.layer.textSymbol.font.family, weight: this.layer.textSymbol.font.weight });
-            textShapes.push(textShape);
-            textShape.rawNode.setAttribute("pointer-events", "none");
-        }
-
-        let rectPadding = 2;
-        let textBox = tooltipGroup.getBoundingBox();
-        let rectShape = tooltipGroup.createRect({ x: textBox.x - rectPadding, y: textBox.y - rectPadding, width: textBox.width + (rectPadding * 2), height: textBox.height + (rectPadding * 2), r: 0 })
-            .setFill([255, 255, 255, 0.9])
-            .setStroke({ color: "#000", width: 0.5 });
-        rectShape.rawNode.setAttribute("pointer-events", "none");
-
-        //handle any view rotation
-        if (this.view.rotation !== 0) {
-            let m = gfx.matrix;
-            let ttc = this.getShapeCenter(rectShape);
-            rectShape.applyTransform(m.rotategAt(360 - this.view.rotation, ttc.x, ttc.y));
-        }
-
-        shape.moveToFront();
-        for (let i = 0, len = textShapes.length; i < len; i++) {
-
-            textShapes[i].moveToFront();
-
-            //handle any view rotation
-            if (this.view.rotation !== 0) {
-                let m = gfx.matrix;
-                let ttc = this.getShapeCenter(textShapes[i]);
-                textShapes[i].applyTransform(m.rotategAt(360 - this.view.rotation, ttc.x, ttc.y));
-            }
-        }
-
-    }
-
-    private destroyTooltip(e) {
-        query(".tooltip-text", this.surface.rawNode).forEach(domConstruct.destroy);
-    }
-
-
-    private removeNodeFromDom(node) {
-        let parent = node.parentNode || node.parentElement;
-        if (parent) parent.removeChild(node);
-
-    }
-
-    private getClusterVectorByGraphicId(id): any {
-        for (var i = 0, len = this.clusterVectors.length; i < len; i++) {
-            if (this.clusterVectors[i].graphic.uid === id) return this.clusterVectors[i];
-        }
-        return undefined;
-    }
-
-    private getClusterVectorFromGroupNode(groupNode): any {
-        for (let v of this.clusterVectors) {
-            if (v.clusterGroup.rawNode === groupNode) {
-                return v;
-            }
-        }
-    }
-
-    private getShapeCenter(shape) {
-        let bbox = shape.getBoundingBox();
-        let x = bbox.x + bbox.width / 2;
-        let y = bbox.y + bbox.height / 2
-        return { x: x, y: y };
-    }
-
-    //IE / Edge don't have the classList property on svg elements, so we can'tuse that add / remove classes - probably why dojo domClass doesn't work either.
-    //so do the following two functions are dodgy string hacks to add / remove classes. Uses a timeout for any css transitions to work correctly.
-    private addClassToNode(node, className, timeoutMs) {
-
-        setTimeout(() => {
-            node.setAttribute("class", node.getAttribute("class") + " " + className);
-        }, timeoutMs); 
-    }
-
-    private removeClassFromNode(node, className, timeoutMs) {
-        let currentClass = node.getAttribute("class");
-        if (!currentClass) return;
-        setTimeout(() => {
-            node.setAttribute("class", currentClass.replace(" " + className, ""));
-        }, timeoutMs);
-    }
-
-}
-
-
-export class FlareClusterLayerExternalRenderer {
-
-    layerView: any;
-    graphics: any;
-
-    parentDiv: any;
-    surface: any;
-
-    activeCluster: any;
-    clusterDeactivatingId: number;
-
-    vectorGroup: VectorGroup;
-
-    get loadedGraphics(): any {
-        return this.layerView.loadedGraphics;
-    }
-
-    constructor(layerView: any) {
-        this.layerView = layerView;
-
-        on.pausable(this.layerView.view.canvas, "mousemove", (e) => this.mouseMove(e));
-
-        this.vectorGroup = new VectorGroup();
-    }
-
-    setup(context) {
-
-        this.parentDiv = this.layerView.view.canvas.parentElement || this.layerView.view.canvas.parentNode;
-
-        this.surface = gfx.createSurface(this.parentDiv, "0", "0");
-        domStyle.set(this.surface.rawNode, { position: "absolute", top: "0", zIndex: -1 });
-        domAttr.set(this.surface.rawNode, "overflow", "visible");
-
-        //init some props of the temp vector group to make
-        this.vectorGroup.surface = this.surface;
-        this.vectorGroup.view = this.layerView.view;
-        this.vectorGroup.view.rotation = 0;
-        this.vectorGroup.transform = [0, 0, 0, 0, 0, 0];
-    }
-
-
-    render(context) {
-        this.graphics = this.layerView.layerViewCore.graphicsCore.graphics;
-
-        let layer = this.layerView.layer;
-
-        //hide the area shapes and flare shapes by default
-        if (this.loadedGraphics.items) {
-            for (let g of this.loadedGraphics.items) {
-
-                if (g.attributes.isFlare || g.attributes.isFlareText) {
-                    g.visible = false;
-                }
-
-                //hide the area unless it's set to always display
-                if (g.attributes.isClusterArea && layer.clusterAreaDisplay !== "always") {
-                    g.visible = false;
-                }
-
-            }
-        }
-
-
-        if (this.activeCluster) {
-            //if a cluster is active and the area graphic should be displayed then make it visible.
-            if (this.activeCluster.areaGraphic && layer.clusterAreaDisplay === "activated") {
-                this.activeCluster.areaGraphic.visible = true;
-            }
-
-            externalRenderers.requestRender(this.layerView.view);
-        }
-
-    }
-
-    private activateCluster(clusterGraphic) {
-
-        if (this.activeCluster !== clusterGraphic) {
-            this.deactivateCluster();
-        }
-        else {
-            return;
-        }
-
-        if (!clusterGraphic) return;
-
-        this.activeCluster = clusterGraphic;
-
-        this.activeCluster.textGraphic = this.activeCluster.attributes.textGraphic;
-        this.activeCluster.areaGraphic = this.activeCluster.attributes.areaGraphic;
-        this.activeCluster.flareGraphics = this.activeCluster.attributes.flareGraphics;
-
-
-        this.setupSurface(this.activeCluster);
-        this.setupClusterGraphic(this.activeCluster);
-        this.createFlares(this.activeCluster);
-
-        externalRenderers.requestRender(this.layerView.view);
-    }
-
-
-    private setupClusterGraphic(clusterGraphic) {
-
-        clusterGraphic.visible = false;
-        clusterGraphic.textGraphic.visible = false;
-
-        //we're going to replicate a cluster graphic in the svg element. Just so it can be styled easily. Again native WebGL would probably be better, but at least this way css can still be used to style/animate things.
-        let sp = this.layerView.view.toScreen(clusterGraphic.geometry);
-        let clusterSymbol = clusterGraphic.symbol;
-        if (!clusterSymbol) {
-            let layer = this.layerView.layer;
-            let info = layer.renderer.getClassBreakInfo(clusterGraphic);
-            clusterSymbol = info.symbol;
-        }
-
-        let clusterGroup = this.surface.createGroup();
-        clusterGroup.rawNode.setAttribute("class", "cluster-group created");
-        clusterGraphic.clusterGroup = clusterGroup;
-
-        //Fake out creation of a dojo shape from graphic object using the temp vectorGroup _drawPoint function
-
-        let webGeo = clusterGraphic.geometry.spatialReference.isWebMercator ? clusterGraphic.geometry : webMercatorUtils.geographicToWebMercator(clusterGraphic.geometry);
-        let tempVector: any = {
-            extent: clusterGraphic.geometry,
-            geometry: clusterGraphic.geometry,
-            graphic: clusterGraphic,
-            symbol: clusterSymbol,
-            projectedGeometry: webGeo
-        };
-
-        //set some default props on the temp vector group object
-        let proj = this.vectorGroup["_projector"];
-        proj["_transformers"] = [];
-        proj["transform"] = [0, 0, 0, 0, 0, 0];
-
-        let clusterShape = this.vectorGroup._drawPoint(this.surface, webGeo, clusterSymbol, tempVector, [0]);
-        clusterShape.setFill(clusterSymbol.color);
-        clusterGroup.add(clusterShape);
-
-        this.addClassToNode(clusterShape.rawNode, "cluster", 5);
-
-        if (clusterSymbol.outline) {
-            clusterShape.setStroke(clusterSymbol.outline.color);
-        }
-
-        let textSymbol = clusterGraphic.textGraphic.symbol;
-        let textSize = textSymbol.font.size * 1.5;
-        let text = clusterGroup.createText({ x: 0, y: textSymbol.font.size / 2, text: textSymbol.text, align: "middle" })
-            .setFill(textSymbol.color)
-            .setFont({ size: textSize, family: textSymbol.font.family, weight: textSymbol.font.weight })
-        text.rawNode.setAttribute("pointer-events", "none");
-
-        this.addClassToNode(text.rawNode, "cluster-text", 5);
-        this.addClassToNode(clusterGroup.rawNode, "activated", 10);
-    }
-
-    private createFlares(clusterGraphic) {
-
-        //flares can only be circles in a scene view for now.
-        let sp = this.layerView.view.toScreen(clusterGraphic.geometry);
-        let clusterGroup = clusterGraphic.clusterGroup;
-        if (!clusterGroup) {
-            clusterGroup = this.surface.createGroup();
-        }
-
-        let radius = 8;
-        let buffer = 5;
-
-        let clusterSymbol = clusterGraphic.symbol;
-        if (!clusterSymbol) {
-            let layer = this.layerView.layer;
-            let info = layer.renderer.getClassBreakInfo(clusterGraphic);
-            clusterSymbol = info.symbol;
-        }
-
-        let bbox = clusterGroup.getBoundingBox();
-        let conCircleRadius = bbox.width / 1.5 + buffer + (radius * 2); //get the radius of the circle to contain everything
-
-        let containerCircle = clusterGroup.createCircle({ cx: 0, cy: 0, r: conCircleRadius })
-            .setFill([0, 0, 0, 0]);
-
-        let flareGraphics = [];
-        for (let i = 0, len = clusterGraphic.flareGraphics.length; i < len; i++) {
-            if (clusterGraphic.flareGraphics[i].attributes.isFlare) {
-                flareGraphics.push(clusterGraphic.flareGraphics[i]);
-            }
-        }
-
-
-        let flareCount = flareGraphics.length;
-        //if there's an even amount of flares, position the first flare to the left, minus 180 from degree to do this.
-        //for an add amount position the first flare on top, -90 to do this. Looks more symmetrical this way.
-        let degreeVariance = (flareCount % 2 === 0) ? -180 : -90;
-
-        //array to hold the animations for displaying flares
-        let stAnims = [];
-
-        for (let i = 0; i < flareCount; i++) {
-
-            let flareGraphic = flareGraphics[i];
-
-            //get the position of the flare to be placed around the container circle.
-            let degree = parseInt(((360 / flareCount) * i).toFixed());
-            degree = degree + degreeVariance;
-            var radian = degree * (Math.PI / 180);
-
-            //calc the center point of the flare
-            let screenPoint = {
-                x: (conCircleRadius - radius - 5) * Math.cos(radian),
-                y: (conCircleRadius - radius - 5) * Math.sin(radian)
-            };
-
-            let layer = this.layerView.layer;
-            //create a group to hold the flare and the text for the flare
-            let flareGroup = clusterGroup.createGroup();
-            let flareCircle = flareGroup.createCircle({ r: radius + 2, cx: screenPoint.x, cy: screenPoint.y })
-                .setFill(layer.flareSymbol.color);
-
-            if (layer.flareSymbol.outline) {
-                flareCircle.setStroke({ width: layer.flareSymbol.outline.width, color: layer.flareSymbol.outline.color });
-            }
-
-            flareGroup.rawNode.setAttribute("class", "flare-group cluster-object");
-            if (flareGraphic.attributes.flareTextGraphic) {
-
-                let textSize = layer.flareTextSymbol.font.size * 1.5;
-                //add a flare text graphic
-                let text = flareGroup.createText({ x: screenPoint.x, y: screenPoint.y + (textSize / 2) - 1, text: flareGraphic.attributes.flareTextGraphic.symbol.text, align: "middle" })
-                    .setFill(layer.flareTextSymbol.color)
-                    .setFont({ size: textSize, family: layer.flareTextSymbol.font.family, weight: layer.flareTextSymbol.font.weight })
-                text.rawNode.setAttribute("pointer-events", "none");
-
-            }
-
-            flareGroup.rawNode.setAttribute("data-tooltip", flareGraphic.attributes.tooltipText);
-            flareGroup.rawNode.setAttribute("data-center-x", screenPoint.x);
-            flareGroup.rawNode.setAttribute("data-center-y", screenPoint.y);
-            flareGroup.isSummaryFlare = flareGraphic.attributes.isSummaryFlare;
-
-            this.addClassToNode(flareGroup.rawNode, "activated", 10);
-
-            flareGroup.mouseEnter = on.pausable(flareGroup.rawNode, "mouseenter", (e) => this.createTooltip(e));
-            flareGroup.mouseLeave = on.pausable(flareGroup.rawNode, "mouseleave", (e) => this.destroyTooltip(e));
-
-        }
-
-    }
-
-    private createTooltip(e) {
-
-        let flareGroupNode = e.gfxTarget ? e.gfxTarget.rawNode : e.target;
-        let shape = flareGroupNode.__gfxObject__;
-
-        this.destroyTooltip(e);
-
-        let tooltipLength = query(".tooltip-text", shape.rawNode).length;
-        if (tooltipLength > 0) {
-            return;
-        }
-
-        //get the text from the data-tooltip attribute of the shape object
-        let text = shape.rawNode.getAttribute("data-tooltip");
-        if (!text) {
-            console.log("no data-tooltip attribute on element");
-            return;
-        }
-
-        //split on /n character that should be in tooltip to signify multiple lines
-        let lines = text.split("\n");
-
-        //read the center positions from the shape, attributes must be set on whatever node is being passed in. Calculating from getboundingBox wasn't working for some reason
-        let xPos = parseInt(shape.rawNode.getAttribute("data-center-x"));
-        //align on top for normal flare, align on bottom for summary flares.
-        let centerY = parseInt(shape.rawNode.getAttribute("data-center-y"));
-        let yPos = !shape.isSummaryFlare ? centerY - 12 : centerY + 17;
-
-        //create a group to hold the tooltip elements
-        let tooltipGroup = shape.createGroup({ x: xPos, y: yPos });
-        tooltipGroup.rawNode.setAttribute("class", "tooltip-text");
-
-        let textShapes = [];
-        for (let i = 0, len = lines.length; i < len; i++) {
-            let textShape = tooltipGroup.createText({ x: xPos, y: yPos + (i * 10), text: lines[i], align: 'middle' })
-                .setFill("#000")
-                .setFont({ size: 10, family: this.layerView.layer.textSymbol.font.family, weight: this.layerView.layer.textSymbol.font.weight });
-            textShapes.push(textShape);
-            textShape.rawNode.setAttribute("pointer-events", "none");
-        }
-
-        let rectPadding = 2;
-        let textBox = tooltipGroup.getBoundingBox();
-        let rectShape = tooltipGroup.createRect({ x: textBox.x - rectPadding, y: textBox.y - rectPadding, width: textBox.width + (rectPadding * 2), height: textBox.height + (rectPadding * 2), r: 0 })
-            .setFill([255, 255, 255, 0.9])
-            .setStroke({ color: "#000", width: 0.5 });
-        rectShape.rawNode.setAttribute("pointer-events", "none");
-
-        shape.moveToFront();
-        for (let i = 0, len = textShapes.length; i < len; i++) {
-            textShapes[i].moveToFront();
-        }
-
-    }
-
-    private destroyTooltip(e) {
-        query(".tooltip-text", this.surface.rawNode).forEach(domConstruct.destroy);
-    }
-
-    private deactivateCluster() {
-        if (!this.activeCluster) return;
-
-        this.activeCluster.visible = true;
-        this.activeCluster.textGraphic.visible = true;
-
-        if (this.activeCluster.attributes.areaGraphic) {
-            this.activeCluster.attributes.areaGraphic.visible = false;
-        }
-
-        this.clearSurface();
-
-        this.clusterDeactivatingId = this.activeCluster.id;
-        this.activeCluster = null;
-
-        externalRenderers.requestRender(this.layerView.view);
-    }
-
-
-    //svg hack to get flares and cluster grpahic to show and be css animation friendly.
-    private setupSurface(activeCluster) {
-        let sp = this.layerView.view.toScreen(activeCluster.geometry);
-        domStyle.set(this.surface.rawNode, { zIndex: 1, overflow: "visible", width: "1px", height: "1px", left: sp.x + "px", top: sp.y + "px" });
-    }
-
-    private clearSurface() {
-        query(">", this.surface.rawNode).forEach(domConstruct.destroy);
-        domStyle.set(this.surface.rawNode, { zIndex: -1 });
-    }
-
-    mouseMove(e) {
-
-        let mousePos = this.getMousePos(e);
-
-        this.layerView.view.hitTest(mousePos).then((response) => {
-            let graphics = response.results;
-
-            if (graphics.length == 0) {
-                this.deactivateCluster();
-                return;
-            }
-
-            let graphicHit = false;
-            graphics.forEach((r) => {
-                if (r.graphic) {
-                    if (r.graphic.attributes.isCluster || r.graphic.attributes.isClusterText) {
-                        let cluster = r.graphic;
-                        if (cluster.attributes.isClusterText) {
-                            //get the cluster graphic from the text 
-                            for (let i = 0, len = this.loadedGraphics.length; i < len; i++) {
-                                if (!this.loadedGraphics.items[i].attributes.isCluster) continue;
-
-                                if (this.loadedGraphics.items[i].attributes.textGraphic === cluster) {
-                                    cluster = this.loadedGraphics[i];
-                                    break;
-                                }
-                            }
-                        }
-                        this.activateCluster(cluster);
-                    }
-
-                    graphicHit = true;
-                    return;
-                }
-            });
-
-            if (!graphicHit) {
-                this.deactivateCluster();
-            }
-        });
-
-    }
-
-    private getMousePos(e) {
-        var rect = this.layerView.view.canvas.getBoundingClientRect();
-        return {
-            x: e.clientX - rect.left,
-            y: e.clientY - rect.top
-        };
-    }
-     
-    //IE / Edge don't have the classList property on svg elements, so we can't use that add / remove classes - probably why dojo domClass doesn't work either.
-    //So the following two functions are dodgy string hacks to add / remove classes. Uses a timeout for any css transitions to work correctly.
-    private addClassToNode(node, className, timeoutMs) {
-
-        setTimeout(() => {
-            node.setAttribute("class", node.getAttribute("class") + " " + className);
-        }, timeoutMs);
-    }
-
-    private removeClassFromNode(node, className, timeoutMs) {
-        let currentClass = node.getAttribute("class");
-        if (!currentClass) return;
-        setTimeout(() => {
-            node.setAttribute("class", currentClass.replace(" " + className, ""));
-        }, timeoutMs);
     }
 
 }
